Move rendering of Document and collection of assets and styles from server/index  into the server/Document/index
